[REF] Split files to keep the standard.
