+++ l10n_br_account_payment_order/models/__init__.py
@@ -7,3 +7,5 @@ from . import account_payment_line
 from . import account_move_line
 from . import account_payment_mode
 from . import res_partner_bank
+from . import account_payment_method
+from . import account_payment_order
